ipl-chatbot-backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.1
pydantic==2.8.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pandas==2.2.2
google-generativeai==0.7.2  # Official Google Gemini SDK